requirements.txt
tensorflow>=2.10
gradio>=3.0
numpy
Pillow